css/css-anchor-position/position-fallback-005.html

<!DOCTYPE html>
<title>Tests padding properties in position fallbacks</title>
<link rel="author" href="mailto:xiaochengh@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-anchor-position-1/#accepted-try-properties">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>

<style>
.cb {
  width: 300px;
  height: 125px;
  position: relative;
}

.anchor {
  position: absolute;
  width: 100px;
  height: 100px;
  background: orange;
  anchor-name: --a;
}

.anchored {
  position: absolute;
  width: 100px;
  height: 100px;
  background: lime;
  top: anchor(--a top);
  position-fallback: --left-then-right;
}

@position-fallback --left-then-right {
  @try {
    right: anchor(--a left);
    padding-right: 15px;
    padding-top: 25px;
  }
  @try {
    left: anchor(--a right);
    padding-left: 25px;
    padding-bottom: 15px;
  }
}

.child {
  width: 100%;
  height: 100%;
  background: blue;
}
</style>

<body onload="checkLayoutForAnchorPos('.target')">

<div class=cb>
  <div class=anchor style="right: 0"></div>
  <!-- Padding affects child layout -->
  <div class="anchored target" style="box-sizing: border-box"
       data-expected-width=100 data-expected-height=100>
    <div class="child target"
         data-offset-x=0 data-offset-y=25
         data-expected-width=85 data-expected-height=75></div>
  </div>
</div>

<div class=cb>
  <div class=anchor style="right: 0"></div>
  <!-- Padding affects self size and child layout -->
  <div class="anchored target"
       data-expected-width=115 data-expected-height=125>
    <div class="child target"
         data-offset-x=0 data-offset-y=25
         data-expected-width=100 data-expected-height=100></div>
  </div>
</div>

<div class=cb>
  <div class=anchor></div>
  <!-- Padding affects child layout -->
  <div class="anchored target" style="box-sizing: border-box"
       data-expected-width=100 data-expected-height=100>
    <div class="child target"
         data-offset-x=25 data-offset-y=0
         data-expected-width=75 data-expected-height=85></div>
  </div>
</div>

<div class=cb>
  <div class=anchor></div>
  <!-- Padding affects self size and child layout -->
  <div class="anchored target"
       data-expected-width=125 data-expected-height=115>
    <div class="child target"
         data-offset-x=25 data-offset-y=0
         data-expected-width=100 data-expected-height=100></div>
  </div>
</div>

</body>
